juniverse.opam should include the dependencies, not the sources
This repository contains a series of packages from the JS monorepo.
To use duniverse, you need to encode the _dependencies_ of those
packages, so that the duniverse tool can vendor the code required
to build them all.

If you run `duniverse init` after this PR, you'll notice fewer
errors, but not zero errors. Some of the dependencies (such as
notty or cryptokit) have not been ported to dune yet, so you will
need to do so and add a fork to
https://github.com/dune-universe/opam-overlays.

Signed-off-by: Anil Madhavapeddy <anil@recoil.org>